requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
importlib_resources==6.4.5
joblib==1.4.2
kiwisolver==1.4.7
numpy==2.0.2
opencv-python==4.10.0.84
packaging==24.1
pdf2image==1.17.0
pillow==11.0.0
pyparsing==3.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
scikit-learn==1.5.2
scipy==1.13.1
six==1.16.0
threadpoolctl==3.5.0
zipp==3.20.2
